requirements2.txt
antlr4-python3-runtime==4.9.2
appdirs==1.4.4
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
crispy-bootstrap5==0.7
dask==2023.9.2
dj-database-url==2.1.0
Django==4.2.3
django-crispy-forms==2.0
djangorestframework==3.14.0
fire==0.5.0
fsspec==2023.9.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
lark-parser==0.12.0
llvmlite==0.41.0
locket==1.0.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.23.3
packaging==23.1
partd==1.4.0
protobuf==3.6.1
psycopg2==2.9.7
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
quantum-blackbird==0.5.0
quantum-xir==0.2.2
requests==2.31.0
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
StrawberryFields==0.23.0
sympy==1.12
termcolor==2.3.0
thewalrus==0.20.0
toml==0.10.2
toolz==0.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
webapp2==2.5.2
xanadu-cloud-client==0.3.1
zipp==3.17.0
